Reorganized files for bonus with diffuse materials
