+++ src/render/render.c
@@ -6,7 +6,7 @@
 /*   By: dloustal <marvin@42.fr>                      +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2025/08/19 13:36:40 by dloustal      #+#    #+#                 */
-/*   Updated: 2025/10/06 14:20:10 by dloustalot    ########   odam.nl         */
+/*   Updated: 2025/10/06 16:48:07 by dloustalot    ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -39,41 +39,6 @@ bool	find_closest_hit(t_ray ray, t_scene *scene, t_object **out_obj,
 	return (true);
 }
 
-static t_color	shade_indirect(t_coord offset_p, t_vec normal, t_scene *scene,
-	int depth, t_ray camera_ray, t_object *obj)
-{
-	t_vec	dir;
-	t_ray	bounce;
-	t_color	recursive_cl;
-
-	dir = sum_vec(normal, random_unit_vec());
-	if (dot(dir, normal) < 0.0)
-		dir = neg_vec(dir);
-	dir = normalized(dir);
-	bounce = set_ray(offset_p, dir);
-	recursive_cl = trace_color(bounce, camera_ray, scene, depth - 1);
-	return (col_scale(recursive_cl, obj->material.albedo));
-}
-
-t_color	trace_color(t_ray ray, t_ray camera_ray, t_scene *scene, int depth)
-{
-	t_object	*obj;
-	double		t_min;
-	t_coord		hit_p;
-	t_vec		normal;
-	t_coord		offset_p;
-
-	if (depth <= 0)
-		return (color(0.0, 0.0, 0.0));
-	if (!find_closest_hit(ray, scene, &obj, &t_min))
-		return (calc_background_color(ray));
-	hit_p = ray_at(ray, t_min);
-	normal = get_normal(obj, hit_p);
-	offset_p = sum_vec(hit_p, scaled(normal, 1e-6));
-	return (col_sum(shade_indirect(offset_p, normal, scene, depth, camera_ray, obj),
-			calc_obj_color(obj, scene, camera_ray, t_min)));
-}
-
 static t_ray	sample_camera_ray(int w, int h, t_vport *vp, t_scene *scene)
 {
 	t_vec	offset;
@@ -89,6 +54,7 @@ static t_ray	sample_camera_ray(int w, int h, t_vport *vp, t_scene *scene)
 	ray = set_ray(scene->camera.pos, sub_vec(pixel_sample, scene->camera.pos));
 	return (ray);
 }
+
 
 void	render_image(mlx_image_t *img, t_scene *scene, t_vport *vp)
 {
